+++ R/assertives.R
@@ -1,29 +1,41 @@
-assertFasttimeOk <- function(.dateTime, .helpers) {
+assertNAstatusPeriodicityOK <- function(
+  na.status,
+  periodicity,
+  level = c("error", "warning")
+) {
+  level <- match.arg(level)
+
+  msg <- paste(
+    "This functionality may only give complete and correct results for time series with explicit missing values and recognised periodicity.",
+    'Consider calling "alter()" with "na.status = \'explicit\'" and/or specified "by" argument first.',
+    sep = "\n"
+  )
+  if (na.status != "explicit" || periodicity == "unrecognised") {
+    if (level == "error") {
+      stop(msg, call. = FALSE)
+    } else {
+      warning(msg, call. = FALSE)
+    }
+  }
+
+  TRUE
+}
+
+assertFasttimeOK <- function(.dateTime, .helpers) {
   if (!requireNamespace("fasttime", quietly = TRUE)) {
-    stop('Package "fasttime" must be installed for this function.', call. = FALSE)
+    stop('Package "fasttime" must be installed for this TALF.', call. = FALSE)
   }
   if (year(.dateTime[1L]) < 1970L || year(last(.dateTime)) > 2199L) {
     stop(
-      "Dates must be between the years 1970 and 2199 for this function.",
+      "Dates must be between the years 1970 and 2199 for this TALF.",
       call. = FALSE
     )
   }
   if (.helpers$timezone != "UTC") {
-    stop('Time zone must be "UTC" for this function.', call. = FALSE)
+    stop('Time zone must be "UTC" for this TALF.', call. = FALSE)
   }
-}
 
-assertRecognisedPeriodicity <- function(periodicity) {
-  if (periodicity == "unrecognised") {
-    stop(
-      paste(
-        "This functionality does not work with time series of unrecognised periodicity.",
-        'Please call "alter()" with specified "by" argument first.',
-        sep = "\n"
-      ),
-      call. = FALSE
-    )
-  }
+  TRUE
 }
 
 assertNoBeginningDot <- function(x) {
@@ -33,4 +45,6 @@ assertNoBeginningDot <- function(x) {
       call. = FALSE
     )
   }
+
+  x
 }